+++ src/CardsAPI/tests/test_hand.py
@@ -2,7 +2,7 @@ from src.CardsAPI import card, Hand
 
 # To launch test in console :
 # ... $ python -m pytest tests/
-from src.CardsAPI.Exceptions import CardsAPIError
+from src.CardsAPI.exceptions import CardsAPIError
 
 
 def test_init_hand():
